Update hello world application
