+++ workmail-hello-world-python/README.md
@@ -1,17 +1,26 @@
 # Amazon WorkMail Hello World
 
-This is a hello world example of the WorkMail Lambda feature. For more information see [AWS WorkMail Lambda documentation](https://docs.aws.amazon.com/workmail/latest/adminguide/lambda.html)
+This is a Hello World function that shows Amazon WorkMail integration 
+with AWS Lambda. When you deploy it to your account, the system creates a 
+Lambda function with all the necessary resources and permissions to access and modify incoming events from WorkMail.
 
-## Setup
+Add your own business logic to the Lambda function based on your use-case.
+For more information see [AWS WorkMail Lambda documentation](https://docs.aws.amazon.com/workmail/latest/adminguide/lambda.html)
+
+## Development
+
+### Using the AWS Lambda Console
 1. Deploy this application via [AWS Serverless Application Repository](https://serverlessrepo.aws.amazon.com/applications/arn:aws:serverlessrepo:us-east-1:489970191081:applications~workmail-hello-world-python)
 2. Open the [WorkMail Console](https://console.aws.amazon.com/workmail/) and create a **RunLambda** [Email Flow Rule](https://docs.aws.amazon.com/workmail/latest/adminguide/create-email-rules.html) that uses this Lambda function.
 
-You now have a working Lambda function that will be triggered by WorkMail based on the rule you created. To add business logic to your Lambda function, open the [AWS Lambda Console](https://eu-west-1.console.aws.amazon.com/lambda/home?region=eu-west-1#/functions) to edit and test your Lambda function with the built-in code editor.
+You now have a working Lambda function that will be triggered by WorkMail based on the rule you created. To add business logic to your Lambda function, open the [AWS Lambda Console](https://us-east-1.console.aws.amazon.com/lambda/home?region=us-east-1#/functions) to edit and test your Lambda function with the built-in code editor.
 
 For more advanced use cases, such as changing your CloudFormation template to create additional AWS resources that will support this application, follow the instructions below.
 
-### Local development
+### Developing locally
 Clone this repository from [GitHub](https://github.com/aws-samples/amazon-workmail-lambda-templates).
+
+We recommend creating and activating a virtual environment, for more information see [Creation of virtual environments](https://docs.python.org/3/library/venv.html).
 
 If you are not familiar with CloudFormation templates, see [Learn Template Basics](https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/gettingstarted.templatebasics.html).
 
@@ -19,37 +28,47 @@ If you are not familiar with CloudFormation templates, see [Learn Template Basic
 2. Modify your Lambda function by changing [app.py](https://github.com/aws-samples/amazon-workmail-lambda-templates/blob/master/workmail-hello-world-python/workmail-hello-world/app.py).
 3. Test your Lambda function locally:
     1. [Set up the SAM CLI](https://aws.amazon.com/serverless/sam/).
-    2. Configure test event at `event.json`.
-    3. Invoke your Lambda function locally using:
-        `sam local invoke WorkMailHelloWorldFunction -e event.json`
+    2. Configure environment variables at `tst/env_vars.json`.
+    3. Configure test event at `tst/event.json`.
+    4. Invoke your Lambda function locally using:
+    
+        `$ sam local invoke WorkMailHelloWorldFunction -e tst/event.json --env-vars tst/env_vars.json`
 
 Once you have validated that your Lambda function behaves as expected, you are ready to deploy this Lambda function.
 
-### Deploying
-This step bundles all your code and configuration to the given S3 bucket.
+### Deployment
+If you develop using the AWS Lambda Console, then this section can be skipped.
+
+Please create an S3 bucket if you do not have one yet, see [How do I create an S3 Bucket?](https://docs.aws.amazon.com/AmazonS3/latest/user-guide/create-bucket.html).
+and check how to create a [Bucket Policy](https://docs.aws.amazon.com/serverlessrepo/latest/devguide/serverlessrepo-how-to-publish.html#publishing-application-through-cli).
+We refer to this bucket as `<Bucket-Name-For-Deployment>`.
+
+This step bundles all your code and configuration to the given S3 bucket. 
 
 ```bash
-sam package \
+$ sam package \
  --template-file template.yaml \
  --output-template-file packaged.yaml \
- --s3-bucket $YOUR_BUCKET_NAME
+ --s3-bucket <Bucket-Name-For-Deployment>
 ```
 
 This step updates your Cloud Formation stack to reflect the changes you made, which will in turn update changes made in the Lambda function.
 ```bash
-sam deploy \
+$ sam deploy \
  --template-file packaged.yaml \
  --stack-name workmail-hello-world \
  --capabilities CAPABILITY_IAM
 ```
 Your Lambda function is now deployed. You can now configure WorkMail to trigger this function.
 
-### Configure WorkMail
-To use this Lambda with WorkMail, follow the instructions at [WorkMail Console](https://console.aws.amazon.com/workmail/) to create a **RunLambda** [Email Flow Rule](https://docs.aws.amazon.com/workmail/latest/adminguide/create-email-rules.html). These steps will require the full ARN of your Lambda function, which can be retrieved by running the following:
+## Frequently Asked Questions
+### Where are the logs?
+You can find the logs in CloudWatch. For more information see [Accessing Amazon CloudWatch logs for AWS Lambda](https://docs.aws.amazon.com/lambda/latest/dg/monitoring-cloudwatchlogs.html).
 
-```bash
-aws cloudformation describe-stacks \
-  --stack-name workmail-hello-world \
-  --query 'Stacks[].Outputs[0].OutputValue'
-```
+### How do I obtain a real message id for my `event.json` file?
+1. Make sure your Lambda function prints out the message id from an event.
+2. Deploy your Lambda function and add the Lambda rule to your WorkMail organization.
+3. Send a test email from your WorkMail account.
+4. Check your CloudWatch logs for a printed message id.
 
+Note that you can access messages in transit for a maximum of one day.
