+++ workmail-hello-world-python/template.yaml
@@ -1,25 +1,106 @@
 AWSTemplateFormatVersion: '2010-09-09'
 Transform: AWS::Serverless-2016-10-31
 Description: >
-    WorkMail Hello World Lambda SAM
-    
-Resources:
-    WorkMailHelloWorldFunction:
-        Type: AWS::Serverless::Function 
-        Properties:
-            CodeUri: workmail-hello-world/
-            Handler: app.lambda_handler
-            Runtime: python3.6
-            Timeout: 10
+  WorkMail Hello World Lambda SAM
 
-    PermissionToCallLambdaAbove:
-        Type: AWS::Lambda::Permission
-        DependsOn: WorkMailHelloWorldFunction
-        Properties: 
-          Action: lambda:InvokeFunction
-          FunctionName: !Ref WorkMailHelloWorldFunction
-          Principal: !Sub 'workmail.${AWS::Region}.amazonaws.com'
+Resources:
+  WorkMailHelloWorldFunction:
+    Type: AWS::Serverless::Function
+    Properties:
+      CodeUri: src/
+      Handler: app.lambda_handler
+      Runtime: python3.7
+      Timeout: 10
+      Role: !GetAtt WorkMailHelloWorldFunctionRole.Arn
+      Layers:
+        - !Sub arn:aws:lambda:${AWS::Region}:489970191081:layer:WorkMailLambdaLayer:2
+      Environment:
+        Variables:
+          UPDATED_EMAIL_S3_BUCKET:
+            Ref: UpdatedEmailS3Bucket
+
+  PermissionToCallLambdaAbove:
+    Type: AWS::Lambda::Permission
+    DependsOn: WorkMailHelloWorldFunction
+    Properties:
+      Action: lambda:InvokeFunction
+      FunctionName: !Ref WorkMailHelloWorldFunction
+      Principal: !Sub 'workmail.${AWS::Region}.amazonaws.com'
+
+  WorkMailHelloWorldFunctionRole:
+    Type: AWS::IAM::Role
+    Properties:
+      AssumeRolePolicyDocument:
+        Statement:
+          - Action:
+              - sts:AssumeRole
+            Effect: Allow
+            Principal:
+              Service:
+                - "lambda.amazonaws.com"
+        Version: "2012-10-17"
+      Path: "/"
+      Policies:
+        - PolicyName: "WorkMailMessageFlowAccessToS3Bucket"
+          PolicyDocument:
+            Version: "2012-10-17"
+            Statement:
+              - Effect: Allow
+                Action:
+                  - "s3:PutObject"
+                Resource: !Sub "${UpdatedEmailS3Bucket.Arn}/*"
+                Condition:
+                  Bool:
+                    aws:SecureTransport:
+                      true
+      ManagedPolicyArns:
+        - "arn:aws:iam::aws:policy/service-role/AWSLambdaBasicExecutionRole"
+        - "arn:aws:iam::aws:policy/AmazonWorkMailMessageFlowFullAccess"
+
+  # WorkMail configures a S3 bucket with all the required policy for reading updated messages.
+  # By default all the object expire after 1 day.
+  # To know more, see https://docs.aws.amazon.com/workmail/latest/adminguide/update-with-lambda.html
+  UpdatedEmailS3Bucket:
+    Type: AWS::S3::Bucket
+    DeletionPolicy: Retain
+    Properties:
+      BucketEncryption:
+        ServerSideEncryptionConfiguration:
+          - ServerSideEncryptionByDefault:
+              SSEAlgorithm: AES256
+      PublicAccessBlockConfiguration:
+        BlockPublicAcls: true
+        BlockPublicPolicy: true
+        IgnorePublicAcls: true
+        RestrictPublicBuckets: true
+      VersioningConfiguration:
+        Status: Enabled
+      LifecycleConfiguration:
+        Rules:
+          - Status: Enabled
+            ExpirationInDays: 1 # Delete after 1 day
+          - Status: Enabled
+            NoncurrentVersionExpirationInDays: 1 # Delete non current versions after 1 day
+
+  UpdatedEmailS3BucketPolicy:
+    Type: AWS::S3::BucketPolicy
+    Properties:
+      Bucket:
+        Ref: UpdatedEmailS3Bucket
+      PolicyDocument:
+        Statement:
+          - Effect: Allow
+            Action:
+              - "s3:GetObject"
+              - "s3:GetObjectVersion"
+            Resource: !Sub "${UpdatedEmailS3Bucket.Arn}/*"
+            Condition:
+              Bool:
+                aws:SecureTransport:
+                  true
+            Principal:
+              Service: !Sub 'workmail.${AWS::Region}.amazonaws.com' # This policy enables WorkMail to read objects from your bucket
 
 Outputs:
-      HelloWorldArn:
-              Value: !GetAtt WorkMailHelloWorldFunction.Arn
+  HelloWorldArn:
+    Value: !GetAtt WorkMailHelloWorldFunction.Arn
