+++ workmail-hello-world-python/tst/env_vars.json
@@ -1,0 +1,5 @@
+{
+  "WorkMailHelloWorldFunction": {
+      "UPDATED_EMAIL_S3_BUCKET": "UPDATED_EMAIL_S3_BUCKET"
+  }
+}
